Cover the module's API with integration tests
Those tests use the previously introduced redis stub server
to send request to a fac simile server, and ensure that the exposed
JS API behaves as intended. We are less intrested in actual redis
behavior correctness than in verifying that the module sends
proper commands to redis, and reacts as expected to somewhat
expected responses.