Add an externally controllable redis test server
In the context of tests, using the `RunT` method will spawn up
a new instance of a stub server speaking redis protocol. The
server's behavior can be observed and controlled using the
`RegisterCommandHandler` method for instance.

The redis stub server has zero dependencies. It is essentially
a stripped down version of miniredis. Most exsting alternatives
as of writing this commit were either pure mocking librarry (not
actually listening on a tcp port), or feature-full servers importing
external dependencies to support things such as Lua that we did not
need.